refactored sc-changelog
