+++ apps/devportal/src/pages/api/changelog/v1/products/index.ts
@@ -1,12 +1,13 @@
 // Interfaces
 import type { NextApiRequest, NextApiResponse } from 'next';
-import { GetProducts } from 'sc-changelog/changelog';
-import Product from 'sc-changelog/types/product';
+import GetProducts from 'sc-changelog/products';
+import { Product } from 'sc-changelog/types';
 import { getQueryValue } from 'sc-changelog/utils/requests';
 
 const handler = async (req: NextApiRequest, res: NextApiResponse<Product[]>) => {
   const showAll: boolean = getQueryValue(req.query.all) == 'false' ? false : true;
   const isPreview = req.preview ? true : false;
+
   res.setHeader('Cache-Control', 'stale-while-revalidate');
 
   await GetProducts(isPreview).then((response: Product[]) => {
