+++ apps/devportal/src/pages/changelog/[product]/index.tsx
@@ -4,8 +4,8 @@ import ChangelogList from '@/src/components/changelog/ChangelogList';
 import Image from 'next/image';
 import Link from 'next/link';
 import { useRouter } from 'next/router';
-import { GetProducts } from 'sc-changelog/changelog';
-import Product from 'sc-changelog/types/product';
+import GetProducts from 'sc-changelog/products';
+import { Product } from 'sc-changelog/types';
 import { slugify } from 'sc-changelog/utils/stringUtils';
 import SmallLinkButton from 'ui/components/buttons/SmallLinkButton';
 import Container from 'ui/components/common/Container';
