+++ apps/devportal/src/pages/changelog/current.tsx
@@ -12,7 +12,7 @@ export default function ChangelogCurrent() {
   return (
     <>
       <Head>
-        <link rel="preload" href="/api/changelog/all?" as="fetch" crossOrigin="anonymous" />
+        <link rel="preload" href="/api/changelog/v1/all?" as="fetch" crossOrigin="anonymous" />
       </Head>
       <Layout title="Sitecore's global changelog" description="Learn more about new versions, changes and improvements">
         <Hero title="Changelog" description="Learn more about new versions, changes and improvements" />
