+++ packages/sc-changelog/changeTypes.ts
@@ -1,0 +1,7 @@
+import { GetAllChangeTypes } from './lib/changeType';
+import { ChangeType, ParseChangeType } from './types';
+
+export default async function GetChangeTypes(isPreview: boolean): Promise<ChangeType[]> {
+  const response = await GetAllChangeTypes(isPreview);
+  return ParseChangeType(response.data);
+}
