+++ packages/sc-changelog/index.ts
@@ -1,2 +1,3 @@
-export * from './configuration';
 export * from './changelog';
+export * from './changetypes';
+export * from './products';
