+++ packages/sc-changelog/types/index.ts
@@ -19,5 +19,6 @@ export enum ProductName {
 
 export * from './changeType';
 export * from './common/media';
+export * from './product';
 export * from './sitecoreCloud';
 export * from './sitecoreProduct';
